requirements.txt
ultralytics
pycocotools
packaging==21.3
easyocr
streamlit
matplotlib
onnxruntime
Pillow
torch
torchvision
numpy
opencv-python
